feat(juxta5-8-prod): uint16 JXV motion; atomic vitals snapshot
Widen vitals motion to uint16_t (cap 65535) in juxta_log and main.
Snapshot motion_events under irq_lock before clearing for vitals.
LIS2DH12: use non-latched INT1 in shared driver. README: non-connectable
adv burst parameter table.

+++ applications/juxta5-8-prod/src/juxta_log.c
@@ -740,7 +740,7 @@ int juxta_log_append_event(struct juxta_log_context *ctx, const struct juxta_set
 	return append_row(ctx, JUXTA_LOG_JXS, row);
 }
 
-int juxta_log_append_vitals(struct juxta_log_context *ctx, uint32_t unix_time, uint8_t motion,
+int juxta_log_append_vitals(struct juxta_log_context *ctx, uint32_t unix_time, uint16_t motion,
 							int32_t batt_mv, int8_t temp_c)
 {
 	static char row[96];
